+++ src/emat/include/emat_thickness_gauge/rviz_emat_panel.h
@@ -5,6 +5,8 @@
 
 #include "emat_thickness_gauge/waveform_widget.h"
 #include "emat/EmatWaveform.h"
+#include "emat/EmatEnvelope.h"
+#include "emat/EmatFeatures.h"
 
 namespace emat {
 
@@ -21,9 +23,13 @@ public:
 
 private:
     void waveformCb(const EmatWaveform::ConstPtr& msg);
+    void envelopeCb(const EmatEnvelope::ConstPtr& msg);
+    void featuresCb(const EmatFeatures::ConstPtr& msg);
 
     WaveformWidget* _widget;
     ros::Subscriber _sub;
+    ros::Subscriber _env_sub;
+    ros::Subscriber _feat_sub;
 };
 
 } // namespace emat