+++ src/emat/src/rviz_emat_panel.cpp
@@ -19,7 +19,9 @@ RvizEmatPanel::RvizEmatPanel(QWidget* parent)
 void RvizEmatPanel::onInitialize() {
     ros::NodeHandle nh;
     _sub = nh.subscribe("/emat/waveform", 10, &RvizEmatPanel::waveformCb, this);
-    ROS_INFO("EMAT Waveform Panel: subscribed to /emat/waveform");
+    _env_sub = nh.subscribe("/emat/envelope", 10, &RvizEmatPanel::envelopeCb, this);
+    _feat_sub = nh.subscribe("/emat/features", 10, &RvizEmatPanel::featuresCb, this);
+    ROS_INFO("EMAT Waveform Panel: subscribed to /emat/waveform, /emat/envelope, /emat/features");
 }
 
 void RvizEmatPanel::waveformCb(const EmatWaveform::ConstPtr& msg) {
@@ -27,6 +29,15 @@ void RvizEmatPanel::waveformCb(const EmatWaveform::ConstPtr& msg) {
     f.raw_data = msg->raw_data;
     f.speed_of_voice = msg->speed_of_voice;
     _widget->pushFrame(f);
+}
+
+void RvizEmatPanel::envelopeCb(const EmatEnvelope::ConstPtr& msg) {
+    std::vector<float> env(msg->envelope.begin(), msg->envelope.end());
+    _widget->pushEnvelope(env, msg->sampling_rate);
+}
+
+void RvizEmatPanel::featuresCb(const EmatFeatures::ConstPtr& msg) {
+    _widget->setThickness(msg->thickness_estimate);
 }
 
 void RvizEmatPanel::save(rviz::Config config) const {